task(#2898): move components examples
