+++ src/examples/hero-banner/HeroBannerExamples.tsx
@@ -1,0 +1,14 @@
+import { HeroBannerWithActions } from "@examples/hero-banner-with-actions.tsx";
+import { SandboxHeader } from "@components/sandbox/sandbox-header/sandboxHeader.tsx";
+
+export const HeroBannerExamples = () => {
+  return (
+    <>
+      <SandboxHeader
+        exampleTitle="Hero Banner with actions"
+        figmaExample="https://www.figma.com/design/aIRjvBzpIUH0GbkffjbL04/%E2%9D%96-Patterns-library-%7C-DDD?node-id=6309-93120&t=X0IQW5flDDaj8Vyg-4">
+      </SandboxHeader>
+      <HeroBannerWithActions />
+    </>
+  );
+};
