+++ src/examples/icon-button/IconButtonExamples.tsx
@@ -1,0 +1,14 @@
+import { ShowMultipleActionsInACompactTable } from "@examples/show-multiple-actions-in-a-compact-table.tsx";
+import { SandboxHeader } from "@components/sandbox/sandbox-header/sandboxHeader.tsx";
+
+export const IconButtonExamples = () => {
+  return (
+    <>
+      <SandboxHeader
+        exampleTitle="Show multiple actions in a compact table"
+        figmaExample="https://www.figma.com/design/aIRjvBzpIUH0GbkffjbL04/%E2%9D%96-Patterns-library-%7C-DDD?node-id=6309-127620&t=X0IQW5flDDaj8Vyg-4">
+      </SandboxHeader>
+      <ShowMultipleActionsInACompactTable />
+    </>
+  );
+};
