+++ src/routes/components/Table.tsx
@@ -1,45 +1,29 @@
-import { useContext, useEffect, useState } from "react";
+import { useState } from "react";
 import { ComponentBinding, Sandbox } from "@components/sandbox";
 import {
   ComponentProperties,
   ComponentProperty,
 } from "@components/component-properties/ComponentProperties";
 import { Category, ComponentHeader } from "@components/component-header/ComponentHeader";
-import {
-  GoabBadge,
-  GoabButton,
-  GoabTab,
-  GoabTable,
-  GoabTableSortHeader,
-  GoabTabs,
-} from "@abgov/react-components";
-import { CodeSnippet } from "@components/code-snippet/CodeSnippet.tsx";
+import { GoabBadge, GoabButton, GoabTab, GoabTable, GoabTabs } from "@abgov/react-components";
 import { GoabTableProps } from "@abgov/react-components";
 import { ComponentContent } from "@components/component-content/ComponentContent";
 import { GoabTableOnSortDetail } from "@abgov/ui-components-common";
-import { LanguageVersionContext } from "@contexts/LanguageVersionContext.tsx";
-import { FilterDataInATable } from "@examples/filter-data-in-a-table.tsx";
 import { omit } from "lodash";
 import { DesignEmpty } from "@components/empty-states/design-empty/DesignEmpty.tsx";
 import { AccessibilityEmpty } from "@components/empty-states/accessibility-empty/AccessibilityEmpty.tsx";
-import { SandboxHeader } from "@components/sandbox/sandbox-header/sandboxHeader.tsx";
-
-interface User {
-  firstName: string;
-  lastName: string;
-  age: number;
-}
+import { TablesExamples } from "@examples/tables/TablesExamples.tsx";
 
 type ComponentPropsType = Omit<GoabTableProps, "onSort"> & {
   onSort?: (sortBy: string, sortDir: number) => void;
 };
 
 export default function TablePage() {
-  const { version } = useContext(LanguageVersionContext);
   const [tableProps, setTableProps] = useState<ComponentPropsType>({
     width: "100%",
   });
-  const FIGMA_LINK = "https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=3785-18038";
+  const FIGMA_LINK =
+    "https://www.figma.com/design/3pb2IK8s2QUqWieH79KdN7/%E2%9D%96-Component-library-%7C-DDD?node-id=3785-18038";
   const [tableBindings, setTableBindings] = useState<ComponentBinding[]>([
     {
       label: "Width",
@@ -122,42 +106,6 @@ export default function TablePage() {
     setTableProps(props as ComponentPropsType);
   }
 
-  // For table demo -- needs to do sort functionality
-  const [users, setUsers] = useState<User[]>([]);
-  useEffect(() => {
-    const _users: User[] = [
-      {
-        firstName: "Christian",
-        lastName: "Batz",
-        age: 18,
-      },
-      {
-        firstName: "Brain",
-        lastName: "Wisozk",
-        age: 19,
-      },
-      {
-        firstName: "Neha",
-        lastName: "Jones",
-        age: 23,
-      },
-      {
-        firstName: "Tristin",
-        lastName: "Buckridge",
-        age: 31,
-      },
-    ];
-    setUsers(_users);
-  }, []);
-
-  function sortData(event: GoabTableOnSortDetail) {
-    const _users = [...users];
-    _users.sort((a: any, b: any) => {
-      return (a[event.sortBy] > b[event.sortBy] ? 1 : -1) * event.sortDir;
-    });
-    setUsers(_users);
-  }
-
   return (
     <>
       <ComponentHeader
@@ -178,7 +126,9 @@ export default function TablePage() {
       <ComponentContent tocCssQuery="goa-tab[open=true] :is(h2[id], h3[id])">
         <GoabTabs initialTab={1}>
           <GoabTab heading="Code playground">
-            <h2 id="component" style={{ display: "none" }}>Playground</h2>
+            <h2 id="component" style={{ display: "none" }}>
+              Playground
+            </h2>
             <Sandbox properties={tableBindings} onChange={onSandboxChange} fullWidth>
               <GoabTable
                 {...omit(tableProps, "onSort")}
@@ -240,394 +190,19 @@ export default function TablePage() {
               </GoabTable>
             </Sandbox>
 
-            <ComponentProperties properties={componentProperties} oldProperties={oldComponentProperties} />
+            <ComponentProperties
+              properties={componentProperties}
+              oldProperties={oldComponentProperties}
+            />
           </GoabTab>
-            <GoabTab
-              heading={
-                <>
-                  Examples
-                  <GoabBadge type="information" content="3" />
-                </>
-              }
-            >
-
-
-              <SandboxHeader
-                exampleTitle="Sort data in a table"
-                figmaExample="https://www.figma.com/design/aIRjvBzpIUH0GbkffjbL04/%E2%9D%96-Patterns-library-%7C-DDD?node-id=6312-97462&t=X0IQW5flDDaj8Vyg-4">
-              </SandboxHeader>
-              <Sandbox fullWidth>
-                <GoabTable onSort={sortData} width="100%">
-                  <thead>
-                    <tr>
-                      <th>
-                        <GoabTableSortHeader name="firstName">First name</GoabTableSortHeader>
-                      </th>
-                      <th>
-                        <GoabTableSortHeader name="lastName">Last name</GoabTableSortHeader>
-                      </th>
-                      <th>
-                        <GoabTableSortHeader name="age" direction="asc">
-                          Age
-                        </GoabTableSortHeader>
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {users.map(user => (
-                      <tr key={user.firstName}>
-                        <td>{user.firstName}</td>
-                        <td>{user.lastName}</td>
-                        <td>{user.age}</td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </GoabTable>
-
-              </Sandbox>
-            {/*React code*/}
-            {version === "old" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="react"
-                allowCopy={true}
-                code={`
-              interface User {
-                firstName: string;
-                lastName: string;
-                age: number;
-              }
-              const [users, setUsers] = useState<User[]>([]);
-              
-              useEffect(() => {
-                const _users: User[] = [
-                  {
-                    firstName: "Christian",
-                    lastName: "Batz",
-                    age: 18
-                  },
-                  {
-                    firstName: "Brain",
-                    lastName: "Wisozk",
-                    age: 19
-                  },
-                  {
-                    firstName: "Neha",
-                    lastName: "Jones",
-                    age: 23
-                  },
-                  {
-                    firstName: "Tristin",
-                    lastName: "Buckridge",
-                    age: 31
-                  }
-                ];
-                setUsers(_users);
-              }, []);
-              
-              function sortData(sortBy: string, sortDir: number) {
-                const _users = [...users];
-                _users.sort((a: any, b: any) => {
-                  return (a[sortBy] > b[sortBy] ? 1 : -1) * sortDir;
-                });
-                setUsers(_users);
-              }
-              
-              return (
-                <GoATable onSort={sortData} width="100%">
-                  <thead>
-                    <tr>
-                      <th>
-                        <GoATableSortHeader name="firstName">First name</GoATableSortHeader>
-                      </th>
-                      <th>
-                        <GoATableSortHeader name="lastName">Last name</GoATableSortHeader>
-                      </th>
-                      <th>
-                        <GoATableSortHeader name="age" direction="asc">Age</GoATableSortHeader>
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                  {users.map(user =>
-                    <tr key={user.firstName}>
-                     <td>{user.firstName}</td>
-                     <td>{user.lastName}</td>
-                     <td>{user.age}</td>
-                    </tr>
-                  )}
-                  </tbody>
-                </GoATable>
-              )
-           `}
-              />
-            )}
-
-            {version === "new" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="react"
-                allowCopy={true}
-                code={`
-              interface User {
-                firstName: string;
-                lastName: string;
-                age: number;
-              }
-              const [users, setUsers] = useState<User[]>([]);
-              
-              useEffect(() => {
-                const _users: User[] = [
-                  {
-                    firstName: "Christian",
-                    lastName: "Batz",
-                    age: 18
-                  },
-                  {
-                    firstName: "Brain",
-                    lastName: "Wisozk",
-                    age: 19
-                  },
-                  {
-                    firstName: "Neha",
-                    lastName: "Jones",
-                    age: 23
-                  },
-                  {
-                    firstName: "Tristin",
-                    lastName: "Buckridge",
-                    age: 31
-                  }
-                ];
-                setUsers(_users);
-              }, []);
-              
-              function sortData(event: GoabTableOnSortDetail) {
-                const _users = [...users];
-                _users.sort((a: any, b: any) => {
-                  return (a[event.sortBy] > b[event.sortBy] ? 1 : -1) * event.sortDir;
-                });
-                setUsers(_users);
-              }
-              
-              return (
-                <GoabTable onSort={sortData} width="100%">
-                  <thead>
-                    <tr>
-                      <th>
-                        <GoabTableSortHeader name="firstName">First name</GoabTableSortHeader>
-                      </th>
-                      <th>
-                        <GoabTableSortHeader name="lastName">Last name</GoabTableSortHeader>
-                      </th>
-                      <th>
-                        <GoabTableSortHeader name="age" direction="asc">Age</GoabTableSortHeader>
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                  {users.map(user =>
-                    <tr key={user.firstName}>
-                     <td>{user.firstName}</td>
-                     <td>{user.lastName}</td>
-                     <td>{user.age}</td>
-                    </tr>
-                  )}
-                  </tbody>
-                </GoabTable>
-              )
-           `}
-              />
-            )}
-
-            {/*Angular code*/}
-            {version === "old" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="angular"
-                allowCopy={true}
-                code={`
-              interface User {
-                firstName: string;
-                lastName: string;
-                age: number;
-              }
-              
-              export class TableComponent() {
-                users: User[] = [];
-                
-                constructor() {
-                  this.users = [
-                    {
-                      firstName: "Christian",
-                      lastName: "Batz",
-                      age: 18
-                    },
-                    {
-                      firstName: "Brain",
-                      lastName: "Wisozk",
-                      age: 19
-                    },
-                    {
-                      firstName: "Neha",
-                      lastName: "Jones",
-                      age: 23
-                    },
-                    {
-                      firstName: "Tristin",
-                      lastName: "Buckridge",
-                      age: 31
-                    }
-                  ];                
-                }
-                
-                handleSort(event: any) {
-                  const {sortBy, sortDir} = event.detail;
-                  this.users.sort(
-                    (a: any, b: any) => (a[sortBy] > b[sortBy] ? 1 : -1) * sortDir
-                  );
-                }
-              }
-              `}
-              />
-            )}
-
-            {version === "old" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="angular"
-                allowCopy={true}
-                code={`
-                <goa-table width="100%" mb="xl" (_sort)="handleSort($event)">
-                  <thead>
-                    <tr>
-                      <th><goa-table-sort-header name="firstName">First name and really long header</goa-table-sort-header></th>
-                      <th><goa-table-sort-header name="lastName">Last name</goa-table-sort-header></th>
-                      <th><goa-table-sort-header name="age" direction="asc">Age</goa-table-sort-header></th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr *ngFor="let user of users; index as i">
-                      <td>{{ user.firstName }}</td>
-                      <td>{{ user.lastName }}</td>
-                      <td>{{ user.age }}</td>
-                    </tr>
-                  </tbody>
-                </goa-table>
-              `}
-              />
-            )}
-
-            {version === "new" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="angular"
-                allowCopy={true}
-                code={`
-              interface User {
-                firstName: string;
-                lastName: string;
-                age: number;
-              }
-              
-              export class TableComponent() {
-                users: User[] = [];
-                
-                constructor() {
-                  this.users = [
-                    {
-                      firstName: "Christian",
-                      lastName: "Batz",
-                      age: 18
-                    },
-                    {
-                      firstName: "Brain",
-                      lastName: "Wisozk",
-                      age: 19
-                    },
-                    {
-                      firstName: "Neha",
-                      lastName: "Jones",
-                      age: 23
-                    },
-                    {
-                      firstName: "Tristin",
-                      lastName: "Buckridge",
-                      age: 31
-                    }
-                  ];                
-                }
-                
-                handleSort(event: GoabTableOnSortDetail) {
-                  const { sortBy, sortDir } = event;
-                  this.users.sort(
-                    (a: any, b: any) => (a[sortBy] > b[sortBy] ? 1 : -1) * sortDir
-                  );
-                }
-              }
-              `}
-              />
-            )}
-
-            {version === "new" && (
-              <CodeSnippet
-                lang="typescript"
-                tags="angular"
-                allowCopy={true}
-                code={`
-                <goab-table width="100%" mb="xl" (onSort)="handleSort($event)">
-                  <thead>
-                    <tr>
-                      <th><goab-table-sort-header name="firstName">First name and really long header</goab-table-sort-header></th>
-                      <th><goab-table-sort-header name="lastName">Last name</goab-table-sort-header></th>
-                      <th><goab-table-sort-header name="age" direction="asc">Age</goab-table-sort-header></th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr *ngFor="let user of users; index as i">
-                      <td>{{ user.firstName }}</td>
-                      <td>{{ user.lastName }}</td>
-                      <td>{{ user.age }}</td>
-                    </tr>
-                  </tbody>
-                </goab-table>
-              `}
-              />
-            )}
-
-              <SandboxHeader
-                exampleTitle="Display numbers in a table so they can be scanned easily"
-                figmaExample="https://www.figma.com/design/aIRjvBzpIUH0GbkffjbL04/%E2%9D%96-Patterns-library-%7C-DDD?node-id=6312-97673&t=X0IQW5flDDaj8Vyg-4">
-              </SandboxHeader>
-            <Sandbox fullWidth>
-              <GoabTable width="100%">
-                <thead>
-                  <tr>
-                    <th>Col 1</th>
-                    <th>Col 2</th>
-                    <th className="goa-table-number-header">Number Column</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td>Item 1</td>
-                    <td>Item 2</td>
-                    <td className="goa-table-number-column">54</td>
-                  </tr>
-                  <tr>
-                    <td>Item 1</td>
-                    <td>Item 2</td>
-                    <td className="goa-table-number-column">4567</td>
-                  </tr>
-                </tbody>
-              </GoabTable>
-            </Sandbox>
-
-              <SandboxHeader
-                exampleTitle="Filter data in a table"
-                figmaExample="https://www.figma.com/design/aIRjvBzpIUH0GbkffjbL04/%E2%9D%96-Patterns-library-%7C-DDD?node-id=7104-1626357&t=WrSJODVw0mryQrrA-4">
-              </SandboxHeader>
-              <FilterDataInATable />
+          <GoabTab
+            heading={
+              <>
+                Examples
+                <GoabBadge type="information" content="3" />
+              </>
+            }>
+            <TablesExamples />
           </GoabTab>
 
           <GoabTab heading="Design">
@@ -636,7 +211,6 @@ export default function TablePage() {
           <GoabTab heading="Accessibility">
             <AccessibilityEmpty figmaLink={FIGMA_LINK} />
           </GoabTab>
-
         </GoabTabs>
       </ComponentContent>
     </>
